create (authorized) folder to require login
